server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
httpx==0.25.1
python-multipart==0.0.6
openai==1.3.5
pillow==10.1.0
pytesseract==0.3.10
gymnasium==0.29.1
numpy==1.26.2
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
google-auth==2.23.4
requests==2.31.0
